fix(draw-route): escape HTML in flow labels on render
Setting the label of a flow to:
    <img src='' onerror='alert(1)'/>
would cause an alert to popup
when the flow is rendered.
This is because the underlying
jsPlumb library calls
.innerHTML for setLabel calls.
This results in a XSS bug.

The label is now santized (escape HTML)
before rendering with this escapeHtml function:
http://stackoverflow.com/a/12034334/2578205

The bug was reported to Mozilla's Bugzilla at:
https://bugzilla.mozilla.org/show_bug.cgi?id=1150658